refactor: extract settings UI into SettingsViews.swift
Move SettingsPanelView and all settings pages/rows (accounts, profile,
identity keys, appearance, privacy/security, audit logs, notifications,
developer mode, relays, key packages) plus the QR-code views out of
MessengerShellView.swift into SettingsViews.swift. Pure code movement:
top-level declarations promoted from file-private to internal;
ProfileImageAvatarView (still in the shell file) likewise promoted so the
settings rows can reference it. No logic or behavior changes.

Build green; test suite unchanged (same two pre-existing image-downsampling
failures, unrelated to view extraction).
